Remove unnecessary style from pre element
Hard coding "white-space:pre;" in the parent element of the results tab-content means long strings like tokens will always exceed the parent box element and force a horizonal scroll bar. This is particularly annoying when you need to copy and use the tokens in other places. It also makes it nearly impossible to use <style> or css changes to turn word wrapping on because the hard coded "white-space:pre;" overwrites everything. Lastly this is the ONLY place where this style is hard coded.